moved teamBuilder to a component
