+++ web-app/src/pages/tournament/TournamentCreator.vue
@@ -6,7 +6,10 @@
       </div>
       <q-form @submit="submitAddTournament">
         <div class="row justify-between">
-          <div class="col-6" style="max-width: 300px">
+          <div
+            :class="$q.screen.gt.xs ? 'col-6' : 'col-12'"
+            :style="$q.screen.gt.xs ? 'max-width: 300px' : ''"
+          >
             <q-input
               clearable
               clear-icon="close"
@@ -18,7 +21,7 @@
               ]"
             />
           </div>
-          <div class="col-6 q-pr-lg">
+          <div class="gt-xs col-6 q-pr-lg">
             <div class="row justify-end">Type of tournament</div>
             <div class="row justify-end">
               <q-radio v-model="tournamentMode" val="type1" label="Type1" />
@@ -30,8 +33,20 @@
               />
             </div>
           </div>
+          <div class="lt-sm col-12 q-pr-lg">
+            <div class="row justify-between items-center">
+              Type of tournament
+              <q-radio v-model="tournamentMode" val="type1" label="Type1" />
+              <q-radio
+                disable
+                v-model="tournamentMode"
+                val="type2"
+                label="Type2"
+              />
+            </div>
+          </div>
         </div>
-        <div class="row justify-between">
+        <div class="gt-xs row justify-between">
           <div class="col-6" style="max-width: 600px">
             <teams-list
               :data="teams"
@@ -40,107 +55,19 @@
             />
           </div>
           <div class="col-6 q-px-md" style="max-width: 600px">
-            <q-form @submit="submitAddTeam">
-              <q-card>
-                <q-card-section>
-                  <div class="text-h6">Team builder</div>
-
-                  <q-input
-                    clearable
-                    clear-icon="close"
-                    outlined
-                    v-model="teamName"
-                    ref="teamNameRef"
-                    label="Team name"
-                    :rules="[
-                      val => (val && val.length > 0) || 'Name cannot be blank!',
-                    ]"
-                  />
-                  <q-select
-                    filled
-                    v-model="player1"
-                    ref="teamMember1"
-                    :options="users"
-                    option-value="val"
-                    option-label="firstName"
-                    label="Captain"
-                    emit-value
-                    map-options
-                    :rules="[
-                      val =>
-                        (val.firstName && val.firstName.length > 0) ||
-                        'Name cannot be blank!',
-                    ]"
-                  >
-                    <template v-slot:no-option>
-                      <q-item>
-                        <q-item-section class="text-grey">
-                          No results
-                        </q-item-section>
-                      </q-item>
-                    </template>
-                    <template v-slot:option="scope">
-                      <q-item v-bind="scope.itemProps" v-on="scope.itemEvents">
-                        <q-item-section avatar>
-                          <q-avatar>
-                            <img :src="scope.opt.avatarSrc" />
-                          </q-avatar>
-                        </q-item-section>
-                        <q-item-section>
-                          {{ scope.opt.firstName }} {{ scope.opt.lastName }}
-                        </q-item-section>
-                      </q-item>
-                    </template>
-                  </q-select>
-                  <q-select
-                    clearable
-                    v-if="player1"
-                    filled
-                    v-model="player2"
-                    ref="teamMember2"
-                    :options="users"
-                    option-value="val"
-                    option-label="firstName"
-                    label="Team member"
-                    emit-value
-                    map-options
-                    :rules="[
-                      val =>
-                        val !== player1 ||
-                        val.length === 0 ||
-                        'Choose another player',
-                    ]"
-                  >
-                    <template v-slot:no-option>
-                      <q-item>
-                        <q-item-section class="text-grey">
-                          No results
-                        </q-item-section>
-                      </q-item>
-                    </template>
-                    <template v-slot:option="scope">
-                      <q-item v-bind="scope.itemProps" v-on="scope.itemEvents">
-                        <q-item-section avatar>
-                          <q-avatar>
-                            <img :src="scope.opt.avatarSrc" />
-                          </q-avatar>
-                        </q-item-section>
-                        <q-item-section>
-                          {{ scope.opt.firstName }} {{ scope.opt.lastName }}
-                        </q-item-section>
-                      </q-item>
-                    </template>
-                  </q-select>
-                  <q-btn
-                    label="Add team"
-                    type="submit"
-                    color="primary"
-                    class="q-mt-md"
-                  />
-                </q-card-section>
-              </q-card>
-            </q-form>
+            <!-- team builder -->
+            <team-builder @team-added="addTeam" :data="users" />
           </div>
+        </div>
+        <div class="lt-sm col-12 q-pt-md">
+          <team-builder @team-added="addTeam" :data="users" />
+
+          <teams-list
+            class="q-pt-md q-pb-md"
+            :data="teams"
+            :columns="columns"
+            :pagination="pagination"
+          />
         </div>
         <q-page-sticky position="bottom-right" :offset="[36, 18]">
           <q-btn color="primary" type="submit" icon="add">
@@ -155,17 +82,18 @@
 <script lang="ts">
 import User from 'src/components/User';
 import { Vue, Component, Ref } from 'vue-property-decorator';
-import teamsList from '../../components/teamsList.vue';
+import teamsList from '../../components/tournament/creator/teamsList.vue';
+import teamBuilder from '../../components/tournament/creator/teamBuilder.vue';
 import { Team } from 'src/components/models';
-import { QInput, QSelect } from 'quasar';
+// import { QInput, QSelect } from 'quasar';
 
 @Component({
-  components: { teamsList },
+  components: { teamsList, teamBuilder },
 })
 export default class TournamentCreator extends Vue {
-  @Ref() readonly teamNameRef!: QInput;
-  @Ref() readonly teamMember1!: QSelect;
-  @Ref() readonly teamMember2!: QSelect;
+  // @Ref() readonly teamNameRef!: QInput;
+  // @Ref() readonly teamMember1!: QSelect;
+  // @Ref() readonly teamMember2!: QSelect;
   private tournamentName = '';
   private tournamentMode = 'type1';
   private pagination = {
@@ -188,10 +116,6 @@ export default class TournamentCreator extends Vue {
       field: 'participants',
     },
   ];
-
-  private teamName = '';
-  private player1: User | null = null;
-  private player2: User | null = null;
 
   private teams: Team[] = [
     {
@@ -235,49 +159,10 @@ export default class TournamentCreator extends Vue {
     console.log('Creating tournament');
     // SENDING THAT CRAP FAAAAR AWAY
   }
-  private submitAddTeam() {
-    const teamToAdd: Team = {
-      name: this.teamName,
-      members: [],
-    };
-    if (this.player1) {
-      teamToAdd.members.push(this.player1);
-    }
-    if (this.player2) {
-      teamToAdd.members.push(this.player2);
-    }
 
-    this.teams.push(teamToAdd);
-
-    // ERROR HERE (players going null)
-    this.teamName = '';
-    this.player1 = null;
-    this.player2 = null;
-
-    // this.teamNameRef.resetValidation();
-    // this.teamMember1.resetValidation();
-    // this.teamMember2.resetValidation();
+  private addTeam(team: Team) {
+    this.teams.push(team);
   }
 }
-
-// methods: {
-//     filterFn (val, update) = {
-//       if (val === '') {
-//         update(() => {
-//           this.options = users.name;
-
-//           // with Quasar v1.7.4+
-//           // here you have access to "ref" which
-//           // is the Vue reference of the QSelect
-//         })
-//         return
-//       }
-
-//       update(() => {
-//         const needle = val.toLowerCase()
-//         this.options = users.name.filter(v => v.toLowerCase().indexOf(needle) > -1)
-//       })
-//     }
-//   }
 </script>
 <style></style>
